Rename `internal` to `util`
